0814 맹희진
휴가 일수 차감

+++ src/main/java/com/groupware/note/user/UserDetails.java
@@ -36,7 +36,7 @@ public class UserDetails {
 	@ManyToOne
 	private Files photo;
 	
-	private long leave;
+	private Integer leave;
 	
 	private long points;
 }
